+++ src/main/java/com/hbm/blocks/generic/BlockOutgas.java
@@ -6,6 +6,7 @@ import com.hbm.handler.radiation.ChunkRadiationManager;
 import net.minecraft.block.Block;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.init.Blocks;
+import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
@@ -16,6 +17,7 @@ public class BlockOutgas extends BlockNTMOre {
     boolean randomTick;
     int rate;
     boolean onNeighbour;
+    private static final int ANCIENT_SCRAP_GAS_RADIUS = 2;
 
     public BlockOutgas(boolean randomTick, int rate, String s) {
         super(s, 1);
@@ -72,6 +74,70 @@ public class BlockOutgas extends BlockNTMOre {
         }
         return Blocks.AIR;
     }
+
+    @Override
+    public void breakBlock(World world, BlockPos pos, IBlockState state) {
+        super.breakBlock(world, pos, state);
+
+        Block block = state.getBlock();
+
+        if (!(block instanceof BlockOutgas outgas)) {
+            return;
+        }
+
+        Block gas = outgas.getGas();
+
+        if (gas == Blocks.AIR) {
+            return;
+        }
+
+        // Spawn gas at broken block location
+        if (isAirBlock(world, pos)) {
+            world.setBlockState(pos, gas.getDefaultState(), 3);
+        }
+
+        // Spawn gas in neighboring blocks if applicable
+        if (outgas.isOnNeighbour()) {
+            spawnGasInAdjacentBlocks(world, pos, gas);
+        }
+
+        // Special handling for ancient scrap - larger radius
+        if (block == ModBlocks.ancient_scrap) {
+            spawnGasInRadius(world, pos, gas, ANCIENT_SCRAP_GAS_RADIUS);
+        }
+    }
+
+    private void spawnGasInAdjacentBlocks(World world, BlockPos pos, Block gas) {
+        for (EnumFacing dir : EnumFacing.values()) {
+            BlockPos adjacentPos = pos.offset(dir);
+            if (isAirBlock(world, adjacentPos)) {
+                world.setBlockState(adjacentPos, gas.getDefaultState(), 3);
+            }
+        }
+    }
+
+    private void spawnGasInRadius(World world, BlockPos center, Block gas, int radius) {
+        for (int x = -radius; x <= radius; x++) {
+            for (int y = -radius; y <= radius; y++) {
+                for (int z = -radius; z <= radius; z++) {
+                    int manhattan = Math.abs(x + y + z);
+
+                    // Skip center and blocks too far away
+                    if (manhattan > 0 && manhattan < 5) {
+                        BlockPos targetPos = center.add(x, y, z);
+                        if (isAirBlock(world, targetPos)) {
+                            world.setBlockState(targetPos, gas.getDefaultState(), 3);
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    private boolean isAirBlock(World world, BlockPos pos) {
+        return world.getBlockState(pos).getBlock() == Blocks.AIR;
+    }
+
 
     @Override
     public void updateTick(World world, BlockPos pos, IBlockState state, Random rand) {
